refactor: consolidate e2e suites under pytest
