+++ package.json
@@ -15,7 +15,7 @@
     "format:check": "prettier --check \"public/js/**/*.js\" \"__tests__/**/*.js\" \"*.js\" \"*.json\" \"*.md\"",
     "lint:format": "npm run lint:fix && npm run format",
     "check": "npm run lint && npm run format:check",
-    "test:e2e": "python scripts/e2e_server.py --server \"python -m http.server 9847 --directory public\" --port 9847 -- python test_run_all.py"
+    "test:e2e": "python -m pytest tests -q"
   },
   "repository": {
     "type": "git",
